remove ref match from CompletionItem
